# apps/backend/src/main/resources/swagger/api.yml
...
openapi: '3.0.1'
info:
  title: 'backend'
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://localhost:8080/api
    description: Development server with TLS Profile
paths:
...
  /v1/grota-financiamentos/dealers:
    get:
      tags:
        - Grota Financiamentos Dealers
      operationId: listDealers
      summary: Listar Lojistas
      description: Retorna uma lista de Lojistas, ordenada por nome (10 por página)
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      responses:
        '200':
          description: Lista de Lojistas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DealerAdminRegisterResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/{id}:
    get:
      tags:
        - Grota Financiamentos Dealers
      operationId: getDealerById
      summary: Obter Lojistas por ID
      description: Retorna os dados de um Lojista com base no ID informado.
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lojista encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerAdminRegisterResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Grota Financiamentos Dealers
      operationId: deleteDealerById
      summary: Deletar logista por ID
      description: Operação de deleção de logistas por ID
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Logista deletado com sucesso
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/{id}/details:
    get:
      tags:
        - Grota Financiamentos Dealers
      operationId: getDealerDetailsById
      summary: Perfil completo do lojista
      description: Retorna o perfil completo do lojista.
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Perfil completo retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerDetailsResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/me/details:
    get:
      tags:
        - Grota Financiamentos Dealers
      operationId: getCurrentDealerDetails
      summary: Perfil completo do lojista autenticado
      description: Retorna o perfil completo do lojista vinculado ao usuário autenticado.
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      responses:
        '200':
          description: Perfil completo retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerDetailsResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/{id}/documents:
    get:
      tags:
        - Grota Financiamentos Dealers
      operationId: getDealerDocuments
      summary: Obter documentos
      description: Retorna todos os documentos do lojista.
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de documentos retornado com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DealerDocumentResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/admin-register:
    post:
      tags:
        - Grota Financiamentos Dealers
      operationId: adminRegisterDealer
      summary: Cadastro de Lojista (admin)
      description: Permite ao admin cadastrar um lojista com endereço completo e dados dos sócios.
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerAdminRegisterRequest'
      responses:
        '201':
          description: Lojista cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerAdminRegisterResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /v1/grota-financiamentos/dealers/profile/update:
    patch:
      tags:
        - Grota Financiamentos Dealers
      operationId: updateDealerProfile
      summary: Atualizar perfil do lojista
      description: Permite ao lojista alterar seus dados de perfil (empresa, CNPJ, endereço, etc).
      security:
        - oauth: [jhipster, email, profile]
        - openId: [jhipster, email, profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerProfileUpdateRequest'
      responses:
        '200':
          description: Perfil do lojista atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerProfileUpdateResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
...
components:
  schemas:
...
    DealerPartnerRequest:
      type: object
      required:
        - cpf
        - name
        - type
        - signatory
      properties:
        cpf:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - SOCIO
        signatory:
          type: boolean
    DealerAdminRegisterRequest:
      type: object
      required:
        - fullName
        - phone
        - enterprise
        - password
        - razaoSocial
        - cnpj
        - partners
      properties:
        fullName:
          type: string
          example: João da Silva
        phone:
          type: string
          example: '11988887777'
        enterprise:
          type: string
          example: Auto Center XPTO
        password:
          type: string
          example: abc123
        razaoSocial:
          type: string
          example: Auto Center XPTO LTDA
        cnpj:
          type: string
          example: '12345678000199'
        address:
          $ref: '#/components/schemas/Address'
        partners:
          type: array
          items:
            $ref: '#/components/schemas/DealerPartnerRequest'
        observation:
          type: string
          example: Cliente indica financiamento de usados.
    DealerAdminRegisterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        razaoSocial:
          type: string
        cnpj:
          type: string
        referenceCode:
          type: string
        phone:
          type: string
        enterprise:
          type: string
        logoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - ATIVO
            - PENDENTE
            - INATIVO
        createdAt:
          type: string
          format: date-time
    DealerProfileUpdateRequest:
      type: object
      required:
        - fullNameEnterprise
        - cnpj
      properties:
        fullNameEnterprise:
          type: string
        birthData:
          type: string
          format: date
        cnpj:
          type: string
        address:
          $ref: '#/components/schemas/Address'
    DealerProfileUpdateResponse:
      type: object
      properties:
        fullNameEnterprise:
          type: string
        birthData:
          type: string
          format: date
        cnpj:
          type: string
        address:
          $ref: '#/components/schemas/Address'
    DealerDetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
        enterprise:
          type: string
        referenceCode:
          type: string
        logoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - ATIVO
            - PENDENTE
            - INATIVO
        fullNameEnterprise:
          type: string
        birthData:
          type: string
          format: date
        cnpj:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
    DealerDocumentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentType:
          type: string
          example: RG_FRENTE
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        reviewStatus:
          type: string
          example: PENDENTE
        reviewComment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Address:
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
...
  responses:
    Problem:
      description: error occurred - see status code and problem object for more information.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the problem type. When dereferenced,
                  it SHOULD provide human-readable documentation for the problem type
                  (e.g., using HTML).
                default: 'about:blank'
                example: 'https://www.jhipster.tech/problem/problem-with-message'
              title:
                type: string
                description: |
                  A short, summary of the problem type. Written in english and readable
                  for engineers (usually not suited for non technical stakeholders and
                  not localized); example: Service Unavailable
              status:
                type: integer
                format: int32
                description: |
                  The HTTP status code generated by the origin server for this occurrence
                  of the problem.
                minimum: 100
                maximum: 600
                exclusiveMaximum: true
                example: 503
              detail:
                type: string
                description: |
                  A human readable explanation specific to this occurrence of the
                  problem.
                example: Connection to database timed out
              instance:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.
  securitySchemes:
    oauth:
      type: oauth2
      description: OAuth2 authentication with KeyCloak
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:9080/realms/jhipster/protocol/openid-connect/auth
          tokenUrl: http://localhost:9080/realms/jhipster/protocol/openid-connect/token
          scopes:
            jhipster: Jhipster specific claims
            email: Email claims
            profile: Profile claims
    openId:
      type: openIdConnect
      description: OpenID Connect authentication with KeyCloak
      openIdConnectUrl: http://localhost:9080/realms/jhipster/.well-known/openid-configuration
security:
  - oauth: [jhipster, email, profile]
  - openId: [jhipster, email, profile]